pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doc-intelligence"
version = "0.1.4"
description = "Document AI - Intelligent document processing and extraction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.57.0",
    "langchain>=1.2.9",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "openai>=2.15.0",
    "pdfplumber>=0.11.9",
    "pypdfium2>=4.30.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.2.0",
    "ollama>=0.6.1",
]

[project.optional-dependencies]
ollama    = ["ollama>=0.4.0"]
anthropic = ["anthropic>=0.40.0"]
gemini    = ["google-genai>=1.57.0"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pyrefly>=0.48.2",
    "ruff>=0.14.13",
    "mkdocs>=1.6.1",
    "pytest>=9.0.2",
    "mkdocs-material>=9.7.1",
    "pre-commit>=4.5.1",
    "pytest-cov>=7.0.0",
    "ipywidgets>=8.1.8",
    "reportlab>=4.4.10",
]

[tool.pytest.ini_options]
markers = [
    "integration: requires live LLM API keys or services",
    "ollama: requires a running Ollama server",
    "openai: requires OPENAI_API_KEY",
    "anthropic: requires ANTHROPIC_API_KEY",
    "gemini: requires GOOGLE_API_KEY",
]

[tool.hatch.build.targets.wheel]
packages = ["doc_intelligence"]

[tool.ruff]
exclude = ["notebooks/"]

[tool.ruff.lint]
select = ["I"]

[tool.pyrefly]
project_excludes = ["**/*.ipynb"]
search_path = ["notebooks"]

[tool.uv.build-backend]
module-name = "doc_intelligence"
module-root = ""
